Fix README formatting issues
This commit addresses the issue of improper rendering of the
README file by correcting the markdown formatting. The changes
include adding spaces after hash symbols for headers and
removing redundant URLs in markdown links. These adjustments
ensure that the markdown is rendered correctly, improving the
readability and presentation of the document.

Fixes #1